tls improvements
