+++ src/Titanium.Web.Proxy/ProxyServer.cs
@@ -246,7 +246,11 @@ public partial class ProxyServer : IDisposable
     /// </summary>
 #pragma warning disable 618
     public SslProtocols SupportedSslProtocols { get; set; } =
-        SslProtocols.Ssl3 | SslProtocols.Tls | SslProtocols.Tls11 | SslProtocols.Tls12;
+        SslProtocols.Ssl3 | SslProtocols.Tls | SslProtocols.Tls11 | SslProtocols.Tls12
+#if NET6_0_OR_GREATER
+        | SslProtocols.Tls13
+#endif
+        ;
 #pragma warning restore 618
 
     /// <summary>
